+++ sickrage/metadata/tivo.py
@@ -27,7 +27,7 @@ import sickrage
 from sickrage.core.exceptions import ShowNotFoundException
 from sickrage.core.helpers import chmodAsParent
 from sickrage.indexers import srIndexerApi
-from sickrage.indexers.indexer_exceptions import indexer_episodenotfound, \
+from sickrage.indexers.exceptions import indexer_episodenotfound, \
     indexer_error, indexer_seasonnotfound, indexer_shownotfound
 from sickrage.metadata import GenericMetadata